eva_common/requirements.txt
# EVA Common shared dependencies
pydantic>=2.0.0
pydantic-settings>=2.0.0
psycopg2-binary>=2.9.0
